+++ packages/cra-template/template/src/Components/ResultsPerPage.tsx
@@ -1,4 +1,4 @@
-import React from 'react';
+import {FunctionComponent, useEffect, useState, useContext} from 'react';
 import Radio from '@material-ui/core/Radio';
 import RadioGroup from '@material-ui/core/RadioGroup';
 import FormControlLabel from '@material-ui/core/FormControlLabel';
@@ -6,50 +6,53 @@ import FormControl from '@material-ui/core/FormControl';
 import Typography from '@material-ui/core/Typography';
 import {
   buildResultsPerPage,
-  ResultsPerPageState,
-  ResultsPerPage as ResultPerPageType,
+  ResultsPerPage as HeadlessResultsPerPage,
 } from '@coveo/headless';
-import {headlessEngine} from '../Engine';
+import EngineContext from '../common/engineContext';
 
-export default class ResultsPerPage extends React.Component {
-  private headlessResultsPerPage: ResultPerPageType;
-  state: ResultsPerPageState;
-
-  constructor(props: any) {
-    super(props);
-
-    this.headlessResultsPerPage = buildResultsPerPage(headlessEngine, {
-      initialState: {numberOfResults: 10},
-    });
-
-    this.state = this.headlessResultsPerPage.state;
-  }
-
-  componentDidMount() {
-    this.headlessResultsPerPage.subscribe(() => this.updateState());
-  }
-
-  updateState() {
-    this.setState(this.headlessResultsPerPage.state);
-  }
-
-  render() {
-    return (
-      <FormControl component="fieldset">
-        <Typography>Results per page</Typography>
-        <RadioGroup
-          row
-          name="test"
-          defaultValue="10"
-          onChange={(event) => {
-            this.headlessResultsPerPage.set(parseInt(event.target.value, 10));
-          }}
-        >
-          <FormControlLabel value="5" control={<Radio />} label="5" />
-          <FormControlLabel value="10" control={<Radio />} label="10" />
-          <FormControlLabel value="15" control={<Radio />} label="15" />
-        </RadioGroup>
-      </FormControl>
-    );
-  }
+interface ResultsPerPageProps {
+  options: number[];
+  controller: HeadlessResultsPerPage;
 }
+const ResultsPerPageRenderer: FunctionComponent<ResultsPerPageProps> = (
+  props
+) => {
+  const {controller, options} = props;
+  const [state, setState] = useState(controller.state);
+
+  useEffect(() => controller.subscribe(() => setState(controller.state)), [
+    controller,
+  ]);
+
+  return (
+    <FormControl component="fieldset">
+      <Typography>Results per page</Typography>
+      <RadioGroup
+        row
+        value={state.numberOfResults}
+        onChange={(event) => {
+          controller.set(parseInt(event.target.value, 10));
+        }}
+      >
+        {options.map((numberOfResults) => (
+          <FormControlLabel
+            key={numberOfResults}
+            value={numberOfResults}
+            control={<Radio />}
+            label={numberOfResults}
+          />
+        ))}
+      </RadioGroup>
+    </FormControl>
+  );
+};
+
+const ResultsPerPage = () => {
+  const engine = useContext(EngineContext)!;
+  const options = [5, 10, 25];
+  const controller = buildResultsPerPage(engine, {
+    initialState: {numberOfResults: options[0]},
+  });
+  return <ResultsPerPageRenderer controller={controller} options={options} />;
+};
+export default ResultsPerPage;